+++ src/ui/DialogWindow.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 /// Begin Dialog Message Member definitions
 
-DialogWindow::DialogWindow(QWidget *parent, const std::string &Title, const std::string &Msg, int status) :
+DialogWindow::DialogWindow(QWidget *parent, const QString &Title, const QString &Msg, int status) :
 QDialog(parent, Qt::WindowTitleHint | Qt::WindowSystemMenuHint | Qt::WindowCloseButtonHint | Qt::MSWindowsFixedSizeDialogHint),
 ui(new Ui::DialogWindow)
 {   
@@ -16,8 +16,8 @@ ui(new Ui::DialogWindow)
     statusIcon = statusIcon.arg(Status::Names[status]);
     ui->_iconLabel->setText(statusIcon);
 
-    this->setWindowTitle(Title.c_str());
-    ui->_Message->setText(Msg.c_str());
+    this->setWindowTitle(Title);
+    ui->_Message->setText(Msg);
 
     connect(ui->_okButton, QPushButton::released, this, accept);
 
@@ -29,12 +29,12 @@ DialogWindow::~DialogWindow()
     delete ui;
 }
 
-QString getTextInputDialog(QWidget *parent, const std::string &Title, const std::string &Message, bool *ok)
+QString getTextInputDialog(QWidget *parent, const QString &Title, const QString &Message, bool *ok)
 {
-    return QInputDialog::getText(parent, Title.c_str(), Message.c_str(), QLineEdit::Normal, "", ok, Qt::WindowTitleHint | Qt::WindowSystemMenuHint | Qt::WindowCloseButtonHint);
+    return QInputDialog::getText(parent, Title, Message, QLineEdit::Normal, "", ok, Qt::WindowTitleHint | Qt::WindowSystemMenuHint | Qt::WindowCloseButtonHint);
 }
 
-QString getItemInputDialog(QWidget *parent, const std::string &Title, const std::string &Message, const QStringList &items, bool *ok)
+QString getItemInputDialog(QWidget *parent, const QString &Title, const QString &Message, const QStringList &items, bool *ok)
 {
-    return QInputDialog::getItem(parent, Title.c_str(), Message.c_str(), items, 0, false, ok, Qt::WindowTitleHint | Qt::WindowSystemMenuHint | Qt::WindowCloseButtonHint);
+    return QInputDialog::getItem(parent, Title, Message, items, 0, false, ok, Qt::WindowTitleHint | Qt::WindowSystemMenuHint | Qt::WindowCloseButtonHint);
 }